CocktailRecommendationService/requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
colorama==0.4.6
psycopg2-binary==2.9.10

torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1

transformers==4.30.0
scikit-learn==1.2.2


konlpy==0.6.0
nltk==3.8.1
nlpaug==1.1.11



tqdm==4.65.0
JPype1==1.4.1
ipywidgets==8.0.0